pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI Boilerplate"
authors = ["Hide <padocon@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.8.1"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.40"}
PyJWT = "^2.4.0"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
celery = "^5.2.7"
gunicorn = "^20.1.0"
fastapi-event = "^0.1.3"
pythondi = "^1.2.4"
ujson = "^5.4.0"
aiomysql = "^0.1.1"
click = "^8.1.3"
greenlet = "1.1.3"
redis = "^4.3.4"
openai = "^0.27.8"
asyncpg = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openai-async = "^0.0.3"
pillow = "^10.0.0"
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.6"
g4f = "^0.0.1.4"
js2py = "^0.74"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
